feat: ADR 0004 — PointXY 内部座標を Float から Double 化 (モデル=Double、描画=Float)
PointXY (_x/_y) と公開 x/y を Double 化し、adjustForRoundingError (1e-5 未満を
0 に潰す Float 時代の getter 補正) を撤去。rotate / lengthXY / magnitude /
calcAngle 系の .toFloat() 落としを除去して Double のまま返す。.toFloat() は
Android Canvas / Compose / Skia 描画 API 直前の縁のみ (規約 B)。Float→Double
の拡幅コンストラクタ 2 本を CSV/UI/紙座標リテラルの縁として追加。CSV/永続化の
数値表現と UI 入力型は不変。CollisionField EPS 1e-2 据え置き (Decision E は
次 phase)。旧 DimOnPath も同時 Double 化し parity 比較対象として存置。

DXF/SFC golden fixture は Double パイプラインの出力で再生成 (orchestrator 裁定:
許容緩和は回帰検出力を恒久的に下げる悪手のため却下。再生成は ADR 0004
Consequences が予告した「より正確な基準への置き直し」の受領形)。

再生成前の差分統計 (旧 Float fixture vs 新 Double fixture、golden テストと
同一 token 規則 + #DATE# 正規化):

  fixture                  差分数値token  最大絶対差(mm)   0/360wrap角度     構造差
  dxf-golden-default            145        0.001376       1 (mod360距離 0)    0
  dxf-golden-dimvariants        212        0.001326       1 (mod360距離 0)    0
  sfc-golden-default            102        0.000689       1 (mod360距離 0)    0
  sfc-golden-dimvariants        165        0.001160       1 (mod360距離 0)    0

全差分は Float ULP (1e4 mm 帯で ~1e-3) 起因の精度向上方向と、角度 0↔360 の
同値 wrap (回転として同一) のみ。非数値 token 差・行数差ゼロ。

検証: :common:desktopTest 172/0、:app:testDevDebugUnitTest 216/0 (1 skipped)、
:desktop:test green、:desktop:build green、DimensionLayoutParityTest (eps=1e-4)
green。

+++ app/src/main/java/com/jpaver/trianglelist/MainActivity.kt
@@ -1566,7 +1566,7 @@ class MainActivity : AppCompatActivity(),
 
     fun flagTriangle(){
         val triangle = trianglelist.get(parameter.number)
-        val tappoint = myview.pressedInModel.scale(PointXY(0f, 0f), 1 / viewscale, -1 / viewscale)
+        val tappoint = myview.pressedInModel.scale(PointXY(0f, 0f), 1.0 / viewscale, -1.0 / viewscale)
         triangle.setPointNumber( tappoint, true )
         setTrianglelist()
     }
@@ -1593,7 +1593,7 @@ class MainActivity : AppCompatActivity(),
             if( !isRotateDedBoxShape ) return
             val current_deduction_number = myview.myDeductionList.lastTapIndex_+1
             val current_deduction = myDeductionList.get(current_deduction_number)
-            current_deduction.rotateShape( current_deduction.point, -degrees )
+            current_deduction.rotateShape( current_deduction.point, (-degrees).toDouble() )
             myview.setDeductionList(myDeductionList, viewscale)
             myview.invalidate()
         }
@@ -2044,7 +2044,7 @@ class MainActivity : AppCompatActivity(),
         PointXY(
             0f,
             0f
-        ), 1f, -1f) }
+        ), 1.0, -1.0) }
 
     fun shadowTapMode(zoomsize: Float): Boolean {
         val RANGE = 0.8f / zoomsize
@@ -2638,7 +2638,7 @@ class MainActivity : AppCompatActivity(),
         )
 
         // translate back by view pointer
-        writer.translate(writer.currentCanvas_, -viewPointer.x, -viewPointer.y)
+        writer.translate(writer.currentCanvas_, -viewPointer.x.toFloat(), -viewPointer.y.toFloat())
         writer.writeTitleFrame(writer.currentCanvas_)
         writer.closeCurrentPage()
 
@@ -2788,8 +2788,8 @@ class MainActivity : AppCompatActivity(),
             // 減算リストのデータをCSVファイルに書き込み
             for (index in 1..myDeductionList.size()) {
                 val dd: Deduction = myDeductionList.get(index)
-                val pointAtRealscale = dd.point.scale(PointXY(0f, 0f), 1 / viewscale, -1 / viewscale)
-                val pointFlagAtRealscale = dd.pointFlag.scale(PointXY(0f, 0f), 1 / viewscale, -1 / viewscale)
+                val pointAtRealscale = dd.point.scale(PointXY(0f, 0f), 1.0 / viewscale, -1.0 / viewscale)
+                val pointFlagAtRealscale = dd.pointFlag.scale(PointXY(0f, 0f), 1.0 / viewscale, -1.0 / viewscale)
 
                 writer.write("Deduction,${dd.num},${dd.name},${dd.lengthX},${dd.lengthY},${dd.overlap_to},${dd.type},${dd.angle},${pointAtRealscale.x},${pointAtRealscale.y},${pointFlagAtRealscale.x},${pointFlagAtRealscale.y},${dd.shapeAngle}")
                 writer.write("\n")
